minor changes to install_zig.sh
- avoid one temporary file with some code golf.
- skip upgrade if the latest version is already installed.

+++ bin/install_zig.sh
@@ -15,21 +15,27 @@ if [[ $OS == "Darwin" ]]; then
     OS="macos"
 fi
 
-curl -sSL -o /tmp/zig-index.json https://ziglang.org/download/index.json
+# without 'echo', 'read' will not read the download URL... for some reason.
+read LATEST_VERSION LATEST_ARCHIVE < <(echo $(curl -sSL https://ziglang.org/download/index.json | \
+                                                  jq --arg os "${ARCH}-${OS}" -r '.master.version, .master[$os].tarball'))
 
-LATEST_VERSION="$(jq -r '.master.version' /tmp/zig-index.json)"
-LATEST_ARCHIVE="$(jq --arg os ${ARCH}-${OS} -r '.master[$os].tarball' /tmp/zig-index.json)"
+if which zig >/dev/null && [[ $(zig version) == $LATEST_VERSION ]]; then
+    echo "Already running the latest version: $LATEST_VERSION"
+    exit 0
+fi
 
 echo "download ${LATEST_VERSION}: $LATEST_ARCHIVE"
 
 curl -# -SL --output-dir /tmp -O "$LATEST_ARCHIVE"
 
+ARCHIVE_FILENAME="/tmp/$(basename $LATEST_ARCHIVE)"
+
 sudo rm -rf /opt/zig
 sudo mkdir /opt/zig
 
-pv "/tmp/zig-${OS}-${ARCH}-${LATEST_VERSION}.tar.xz" | sudo tar -C /opt/zig --strip-components 1 -xzf -
+pv "$ARCHIVE_FILENAME" | sudo tar -C /opt/zig --strip-components 1 -xzf -
 
-rm -f "/tmp/zig-index.json" "/tmp/zig-${OS}-${ARCH}-${LATEST_VERSION}.tar.xz"
+rm -f "$ARCHIVE_FILENAME"
 
 if [[ "$OS" == "darwin" ]]; then
     echo "Setting up /etc/paths.d"
